ProgressBar refactored
