+++ VultisigApp/VultisigApp/iOS/View/Components/VultisigLogo+iOS.swift
@@ -12,5 +12,10 @@ extension VultisigLogo {
     var container: some View {
         content
     }
+    
+    var descriptionContainer: some View {
+        descriptionContent
+            .padding(.top, 10)
+    }
 }
 #endif
